fix encoder glitch issue
we had some noise-glitch on the encoder signal at certain frequencies. The issue come from when we send a new "tick" value over the SPI bus only half the byte was updated.
This was solved by not using the same buffer in the input capture function as for the SPI transfer. Now we are simply copying the values into another bugger before sending them.

We also updated the system clock, added a filet of 0xF on the input and we are now looking at rising insted o falling edges.